% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdjustedRandIndex.R
\name{CalculateARI}
\alias{CalculateARI}
\title{Calculate adjusted rand index}
\usage{
CalculateARI(true_lab, pre_lab)
}
\arguments{
\item{true_lab}{the label of the truth;}

\item{pre_lab}{the label of prediction;}
}
\description{
Calculate adjusted rand index
}
\author{
Chong Chen, \email{cheung1990@126.com}
}
