% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd.R
\name{rmd}
\alias{rmd}
\title{imputation by robust matrix decomposition}
\usage{
rmd(Y, tau = NULL, lambda = NULL, initL = NULL, initS = NULL,
  initLambda = NULL, maxiter = 100, abstol = 0.001, reltol = 0.001,
  rho = 1, overrelax = 1.5, candidate = 0.05, econ = 1)
}
\arguments{
\item{Y}{A single cell RNA data matrix;
rows representing cells.}

\item{tau}{Tuning parameter to penalize the sparsity of S;}

\item{lambda}{Tuning parameter to penalize the row rank of L;}

\item{initL}{The initionlization of L;}

\item{initS}{The initionlization of S;}

\item{initLambda}{The initionlization of Lambda;}

\item{maxiter}{maxmium iteration of algorithm;}

\item{candidate}{the cutoff for candidate drop out;}
}
\description{
imputation by robust matrix decomposition
}
\author{
Chong Chen, \email{cheung1990@126.com}
}
