\name{scRMD-package}
\alias{scRMD-package}
\alias{scRMD}
\docType{package}
\title{
\packageTitle{scRMD}
}
\description{
\packageDescription{scRMD}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{scRMD}
\packageIndices{scRMD}
~~ An overview of how to use the package, including the ~~
~~ most important functions ~~
}
\author{
\packageAuthor{scRMD}

Maintainer: \packageMaintainer{scRMD}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from ~~
~~ file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
